refactor: move static elements to wrapPageElement
Moves the header, footer, and some other shared things to
`wrapPageElement.js`. This means they won't be constantly rerendered on
route change.

Also restores the page transition element, since with this change
transitions don't break the header.

Also adds a small entry animation to the header, since there's a SSR
issue where the desktop header flashes on mobile devices because it's
server rendered. Now it fades in, and won't be shown on mobile.